Remove Microsoft.Extensions.AI.Ollama
