--- src/Libraries/Microsoft.Extensions.AI.Ollama/Microsoft.Extensions.AI.Ollama.csproj
@@ -1,44 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <RootNamespace>Microsoft.Extensions.AI</RootNamespace>
-    <Description>Implementation of generative AI abstractions for Ollama. This package is deprecated, and the OllamaSharp package is recommended.</Description>
-    <Workstream>AI</Workstream>
-  </PropertyGroup>
-
-  <PropertyGroup>
-    <Stage>preview</Stage>
-    <EnablePackageValidation>false</EnablePackageValidation>
-    <MinCodeCoverage>78</MinCodeCoverage>
-    <MinMutationScore>0</MinMutationScore>
-  </PropertyGroup>
-
-  <PropertyGroup>
-    <TargetFrameworks>$(TargetFrameworks);netstandard2.0</TargetFrameworks>
-    <NoWarn>$(NoWarn);CA2227;SA1316;S1121;EA0002</NoWarn>
-    <TreatWarningsAsErrors>true</TreatWarningsAsErrors>
-    <DisableNETStandardCompatErrors>true</DisableNETStandardCompatErrors>
-  </PropertyGroup>
-
-  <PropertyGroup>
-    <InjectCompilerFeatureRequiredOnLegacy>true</InjectCompilerFeatureRequiredOnLegacy>
-    <InjectRequiredMemberOnLegacy>true</InjectRequiredMemberOnLegacy>
-    <InjectSharedDiagnosticIds>true</InjectSharedDiagnosticIds>
-    <InjectSharedEmptyCollections>true</InjectSharedEmptyCollections>
-    <InjectStringHashOnLegacy>true</InjectStringHashOnLegacy>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="System.Net.Http.Json" />
-    <PackageReference Include="System.Text.Json" />
-  </ItemGroup>
-
-  <ItemGroup Condition="'$(TargetFramework)' == 'net462'">
-    <Reference Include="System.Net.Http" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="../Microsoft.Extensions.AI.Abstractions/Microsoft.Extensions.AI.Abstractions.csproj" />
-  </ItemGroup>
-
-</Project>
