--- src/Libraries/Microsoft.Extensions.AI.Ollama/README.md
@@ -1,11 +1,0 @@
-# Microsoft.Extensions.AI.Ollama
-
-This package is deprecated and the [OllamaSharp](https://www.nuget.org/packages/ollamasharp) package is recommended. `OllamaSharp` provides .NET bindings for the [Ollama API](https://github.com/jmorganca/ollama/blob/main/docs/api.md), simplifying interactions with Ollama both locally and remotely.
-
-No further updates, features, or fixes are planned for the `Microsoft.Extensions.AI.Ollama` package.
-
-## Related packages
-
-* [OllamaSharp](https://www.nuget.org/packages/OllamaSharp)
-* [Microsoft.Extensions.AI](https://www.nuget.org/packages/Microsoft.Extensions.AI)
-* [Microsoft.Extensions.AI.Abstractions](https://www.nuget.org/packages/Microsoft.Extensions.AI.Abstractions)
