--- test/Libraries/Microsoft.Extensions.AI.Ollama.Tests/TestJsonSerializerContext.cs
@@ -1,12 +1,0 @@
-﻿// Licensed to the .NET Foundation under one or more agreements.
-// The .NET Foundation licenses this file to you under the MIT license.
-
-using System.Collections.Generic;
-using System.Text.Json.Serialization;
-
-namespace Microsoft.Extensions.AI;
-
-[JsonSerializable(typeof(string))]
-[JsonSerializable(typeof(int))]
-[JsonSerializable(typeof(IDictionary<string, object?>))]
-internal sealed partial class TestJsonSerializerContext : JsonSerializerContext;
